move MXBeans to Measurements, to improve iOS porting
